badge: remove compilation warnings
The following warnings occurred during compilation:
.pio\libdeps\conference\fmt\include/fmt/core.h:3119:44:   required from 'std::string fmt::v8::format(format_string<T ...>, T&& ...) [with T = {nsec::button::id&, nsec::button::event&}; std::string = std::__cxx11::basic_string<char>; format_string<T ...> = basic_format_string<char, nsec::button::id&, nsec::button::event&>]'
components/utils/logging.hpp:46:31:   required from 'void nsec::logging::logger::log(severity, fmt::v8::format_string<T ...>, Args&& ...) const [with Args = {nsec::button::id&, nsec::button::event&}; fmt::v8::format_string<T ...> = fmt::v8::basic_format_string<char, nsec::button::id&, nsec::button::event&>]'
components/utils/logging.hpp:58:12:   required from 'void nsec::logging::logger::info(fmt::v8::format_string<T ...>, Args&& ...) const [with Args = {nsec::button::id&, nsec::button::event&}; fmt::v8::format_string<T ...> = fmt::v8::basic_format_string<char, nsec::button::id&, nsec::button::event&>]'
components\badge\badge.cpp:147:17:   required from here
.pio\libdeps\conference\fmt\include/fmt/core.h:1706:15: warning: possibly dangling reference to a temporary [-Wdangling-reference]

Remove warning by adding typecast when sending "button_callback_context" to the logger.

+++ components/badge/badge.cpp
@@ -144,7 +144,7 @@ uint8_t nr::badge::level() const noexcept
 void nr::badge::on_button_event(nsec::button::id button,
                                 nsec::button::event event) noexcept
 {
-    _logger.info("Button event: button={}, event={}", button, event);
+    _logger.info("Button event: button={}, event={}", (int)button, (int)event);
 
     // Send the received event to the LEDs function.
     _update_leds(button, event);
